demos/Win: tri/cube create a console window if layers enabled.
On MS Windows (i.e. _WIN32), if layers appear to be enabled (i.e. if the
VK_LAYER_NAMES registry value and/or environment variable string(s) aren't
empty), a console window is created.  The demo application won't really know if
a layer will be found and successfully loaded.  However, it can infer that one
or more layers were requested by a non-empty registry value and/or environment
variable (from code copied from the loader).

After the user tries to close the rendering window (e.g. by pressing the red
"X" in the upper-right corner, it is closed by the app, and rendering ceases to
it.  At that time, a string is printed at the end of the console output,
informing the user that they should close the console window after its output
is no longer desired.