intel: add intel_render_pass_{attachment,subpass}
Make intel_render_pass own an array of intel_render_pass_attachment and an
array of intel_render_pass_subpass.  Every subpass references the attachment
array for its outputs.

It gets a bit more clear that intel_fb is for intel_render_pass just like
intel_desc_set is for intel_desc_layout.  It has always been so at the API
level, but it was harder to see.

This is in preparation for an upcoming header change.