tri/cube: Correct WSI swapChainPresentMode fallback is FIFO.
A copy and paste error from the spec made it seem that all ICDs will support
VK_PRESENT_MODE_IMMEDIATE_WSI (a.k.a. "immediate").  That's not true.  They all
need to support VK_PRESENT_MODE_FIFO_WSI.  Changed the comment and code so that
the preference of these demos is:

1) VK_PRESENT_MODE_MAILBOX_WSI
2) VK_PRESENT_MODE_IMMEDIATE_WSI
3) VK_PRESENT_MODE_FIFO_WSI