intel: Wait for xcb present in vkAcquireNextImageWSI
If vkAcquireNextImageWSI doesn't find an unused image, then wait for all
present events that are expected.  Pass the timeout on to x11_swap_chain_wait.
This allows presented images to become displayed and displayed images to become unused.
This still doesn't use a true semaphore for the caller to wait on.