loader/layers: Support unknown phys device exts
Add support to handle unknown physical device extensions to both the
loader and all our LVL layers.  The functionality expects that at least
one layer or ICD supports the unknown extension, and then loader will
return a generic trampoline to the application.  It will also force the
layers to pass through a generic terminator before proceeding to the ICD
(if one supports the extension).

For a layer to support this, it must export the new
vkNegotiateLoaderLayerInterace function at version 2 or newer.  This also
includes supporting and returning the internal
vk_layerGetPhysicalDeviceProcAddr call.

For an ICD to support this, it must export version 4 of the loader/icd
interface as well as export the vk_icdGetPhysicalDeviceProcAddr function.

Change-Id: I1f659fef40ac4343af02ef88fc7584bf45777ea0