loader: Fix early-out in extension
The function extensions_create_instance() attempts
to parse through the list of enabled instance
extensions and mark those known to the loader as
enabled in some internal data structures.  These
enable bits are later used to avoid returning
pointers to instance functions for instance
extensions which are not enabled.

Prior to this change, the function would return
after finding a known extension in the list of
extension names.  This is undersirable, as all
known extensions in the array need to be marked
as enabled.  This change switches the logic to
continue the loop after finding a known extension.