tests:Add DrawWithBufferWaRandRaWConflicts
This is a race condition test to check both a RaW and WaR conflict with
buffers used in a Draw. A src buffer is copied to a dst buffer, then,
without a barrier, the src is used as a storage buffer in a draw and
the dst is used as a uniform buffer in the same draw. This results in a
RaW conflict for the Dst/Uniform buffer and WaR conflict for the
Src/Storage buffer.