layers:Streamline descriptor mem access update
Put memory access updates for active descriptors directly into the read
and write vectors.
This kills an extra data structure transformation where we moved the
accesses through sets before putting them into vectors. There was no
benefit to the extra overhead and this update brings the performance of
validation with memory access checks on-par with validation when the
checks are disabled.

Still have the warning callback disabled for now.

The ENABLE_MEMORY_ACCESS_CALLBACK define in vk_layer_config.h can be
un-commented to enable the warning callback, as well as the descriptor
memory access tracking.