layers:Refactor CmdBeginRenderPass()
Only perform state updates for CmdBeginRenderPass() if we're not
skipping the call to the driver.

Also update validateSubpassCompatibility() to return early if skip is
set to "true." That function will only return a single error code so in
the event that we hit a skip case go ahead and just return early as we
don't need to flag same code multiple times.