layers: PreRecord Free/Destroy calls
To avoid Destroy/Create and Free/Alloc race conditions (when a driver
returns a destroyed(freed) handle between the PreCall and PostCall
phases of a validation intercept routine, PostCallRecord functionality
has been moved to PreCall.  This affects the following calls

	FreeMemory
	DestroyFence
	DestroySemaphore
	DestroyEvent
	DestroyQueryPool
	DestroyBuffer
	DestroyBufferView
	DestroyImage
	DestroyImageView
	DestroyPipeline
	DestroySampler
	DestroyDescriptorSetLayout
	DestroyFramebuffer
	DestroyRenderPass
	FreeDescriptorSets

Did not alter the destroy functions for Instance, Device, or the Debug
or layer extensions.

Change-Id: I5133982eaa1a0c94283fa922d9c3bcc90b5a6223